# perf: optimize meta/metrics dict iteration