@brettlangdon 2021-09-03
Just realizing those performance results are showing `many-tags` as slower, which is the desired outcome.