@brettlangdon 2021-09-03
Yup, ran a few more times. Not faster... this is what I get for trying to outsmart Cython.